requirements.txt
python-dotenv
langchain-google-genai
langchain-core
gradio